requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-samesite-none==0.0.3
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.2.0
django_rest_passwordreset==1.3.0
torch
facenet-pytorch
Pillow
opencv-python
numpy
json

